moving everything into pkg dir
